Fix drawing to incorporate translation
Not sure of the performance hits here. Also seems as though output_width
and char_width are not strictly necessary anymore since we calculate
them as we go along.

Also changed is that when a wide character gets split by a translate or
a region cutoff, the extra space is replaced with '…'. This is necessary
at least on the left to avoid jumps when the character appears when
scrolling. Not sure if this behavior is also wanted on the right. Is
there a "partial character" codepoint for when there isn't enough space
to display a full character?