Optimize Input.Classify.classify.
So I had a problem that classify.compile is evaluated on every single
keystroke. I can only guess that when classify had a call to compile
under a guard, it prevented GHC to optimize this as expected. Switching
'compile table' and pattern matching on input seems to help, I no longer
see it dominating in profiling data.