Resolve #76. Terminfo "clear to end of line" does not mean what I thought it means. I though this would clear the text and current display attributes from the cursor to end of the line. This is not the case. The display attribute is state of the output cursor and not the already dislayed text.
This change sets the default display attributes at the end of a line as well as the start of a new
line.

This passes all the tests but will cause a regression in Yi: The display of a selection area will
not extend to the end of the screen. However, I have a branch of Yi that adds vertical splits. In
order to implement this the selection area display code needed to handle extending the image to the
end of the frame.