Add support for bracketed paste mode in XtermColor-driven terminals (fixes #50)
For details, see http://cirw.in/blog/bracketed-paste

This change adds a new Event constructor,

  EvPaste String

to deliver the contents of a paste to the application.  The constructor
provides a String instead of Text as originally suggested because I
wanted to punt on encoding issues and let the application deal with
that.

This change has been tested under tmux, Terminal.app, and iTerm2. A
development build of GNU screen also supports this paste mode (provided
you set TERM to "xterm" to force Vty to use the XtermColor driver for
screen). The mode has apparently been supported for a few years, but
I couldn't get the latest stable build in Homebrew (4.03.01) to work
in this mode so my only testing with GNU Screen was with git HEAD
(8d677e73).  It's possible the original patch didn't actually work but
interim refactorings fixed it.