Bugfix: make "refresh" actually behave as advertised (fixes #104)
This change fixes "refresh" by clearing the AssumedState of the
DisplayContext used to do the refresh operation.  Before this change,
"refresh" would trigger a redraw but the redraw would check the new
output picture against the previous one and determine that there were no
differences, resulting in a no-op redraw.  This change fixes that
behavior by first clearing the previous draw state of the DisplayContext
to force the output routine to conclude that the new picture is
different on every line.