translateX/Y: fix negative translations
This change makes it so that translating an image negatively will return
an empty image if the negative translation completely hides the input
image. Previously we'd just return an image with a negative height,
which not only violates the requirement that all images have height no
less than zero, but it caused downstream bugs in mutable vector
accesses due to negative vector indices.