Config: support specifying character width map paths
This change adds a new Config record field, termWidthMaps, which holds a
list of pairs that map terminal names to width table file paths.

This change also extends the configuration parser to support lines of
the form

  widthMap <Terminal> <Path>

where <Terminal> must be a double-quoted string literal indicating the
value of TERM for which the specified map should be active, and <Path>
must be a double-quoted string literal specifying the file path to the
width table file to use. E.g.,

  widthMap "xterm" "/home/user/.vty/width_table_xterm.dat"