Reinstate support for input loop event logging
This change adds a field to Input for logging to support the vty-unix
implementation that previously depended on _inputDebug (a Handle) from
Input. This interface is a bit more abstract to allow for platform
differences. This is not intended to be a "good" or principled change to
support logging since that would require broader thinking about the
library as a whole; instead this is aimed merely at reinstating what was
broken by the refactoring.