Reinstate Vty configuration, but under a new name and with structural changes
This patch reinstates the Graphics.Vty.Config module that was removed in
a previous commit as part of migration code to vty-unix. But this
version of the configuration machinery has some important changes:

* The Config type was renamed to VtyUserConfig to better emphasize the
  nature of the data type.

* The VtyUserConfig (previously Config) had all fields that were
  specific to Unix terminals removed, since those belong in vty-unix,
  and also because they were not actually part of the user configuration
  and were more runtime configuration than user configuration. (The fact
  that Config had both user and runtime configuration made the internal
  API messy, and this change is part of cleaning that up.)

This change assumes that the fields left behind in VtyUserConfig are
platform-agnostic, such as the list of unicode table width map files, or
the custom input mappings. If that turns out to be untrue in the long
run, more things may be moved out of VtyUserConfig. But for now those
are left behind, and this change results in the VtyUserConfig storing
only what the config parser was looking for anyway.