mk_wcwidth.c: make custom table lookups fall back to the built-in table
Prior to this change, if a custom character width table was installed,
then that table would have all of its entries initialized to 1. In the
case where someone installed a partial table -- a table that didn't span
the entire character range supported by the table -- then table lookups
for ranges outside that table would yield a width of 1 column. That
isn't a problem when Vty's built-in table generation tool is used since
it probes the entire character range. However, it was a problem for very
small custom tables: if a very small custom table was installed -- one
that doesn't cover the whole character range -- then it would give a
result of 1 for many characters that would get a result of 2 from the
built-in table. That meant that such small custom tables would actually
yield worse results than complete ones and possibly worse results than
just using the built-in table.

This fixes that by making it so that custom table lookup failures always
fall back to the built-in table. This works by initializing the custom
table with a sentinel value and then falling back to the built-in table
when that sentinel value is present for a particular character lookup.

This is assuming that the built-in table is less wrong than yielding a
default of '1' for those failed lookups, and in any case if they are
indeed wrong, then the user can simply update their table to set the
desired width for the character in question.